booking screen dynamic
