+++ src/Screens/Booking/Booking.jsx
@@ -2,86 +2,9 @@ import React, { useState } from "react";
 import { Col, Container, Row, Button, Form } from "react-bootstrap";
 import "./Booking.scss";
 import { LinkContainer } from "react-router-bootstrap";
-import useCounter from "./useCounter";
+import ClassDetails from "./ClassDetails";
 
-const ClassDetails = ({
-  seatClass,
-  price,
-  seatsCount,
-  availableSeats,
-  bookedSeats,
-  updateTicketCounts,
-}) => {
-  const { count, increment, decrement } = useCounter();
-  let available = 25;
-  return (
-    <Container fluid="true" className="count-container">
-      <Row>
-        <Col md="4" sm="6" xs="6" className="seat-class">
-          <h2>{seatClass}</h2>
-        </Col>
-        <Col md="2" sm="6" xs="6">
-          <Row className="seat-cnt">{seatsCount} seats</Row>
-          <Row className="tkt-price">{price} LKR</Row>
-        </Col>
-        <Col md="4" sm="6" xs="6">
-          <Row className="seat-grey">
-            <Col xs={6}>
-              <p>Available:</p>
-            </Col>
-            <Col xs={6}>
-              <p className="avai-para">{availableSeats - count}</p>
-            </Col>
-          </Row>
-          <Row className="seat-grey">
-            <Col xs={6}>Booked:</Col>
-            <Col xs={6} className="booked-para">
-              {seatsCount - availableSeats}
-            </Col>
-          </Row>
-        </Col>
-        <Col
-          md="2"
-          sm="6"
-          xs="6"
-          className="d-flex justify-content-center align-items-center"
-        >
-          <Col md="6" sm="6" xs="6">
-            <span className="count-span" style={{ marginTop: "-30px" }}>
-              {count}
-            </span>
-          </Col>
-          <Col md="6" sm="6" xs="6">
-            <Row>
-              <Button
-                variant="primary"
-                className="btn-inc"
-                onClick={() => {
-                  increment();
-                  updateTicketCounts(seatClass, count + 1);
-                }}
-              >
-                +
-              </Button>
-            </Row>
-            <Row>
-              <Button
-                variant="primary"
-                className="btn-dec"
-                onClick={() => {
-                  decrement();
-                  updateTicketCounts(seatClass, count - 1);
-                }}
-              >
-                -
-              </Button>
-            </Row>
-          </Col>
-        </Col>
-      </Row>
-    </Container>
-  );
-};
+
 
 const NewSchedule = () => {
   const crowdColors = [
@@ -94,7 +17,26 @@ const NewSchedule = () => {
     "#ff9800",
   ];
   const [ticketCounts, setTicketCounts] = useState({ "": 0 });
+  const [totalCost, setTotalCost] = useState(0);
+  
 
+  //update total cost according to ticket count
+  const updateCost=(ticketCost,newCount,count)=>{
+    
+     
+      setTotalCost((prevTotalCost) => {
+        if (newCount < 0) {
+          return prevTotalCost;
+        }
+        else if (newCount>count){
+          return prevTotalCost + ticketCost;
+        }
+        else{
+          return prevTotalCost - ticketCost;
+        }
+      });
+   
+  }
   // Update ticket counts
   const updateTicketCounts = (ticketClass, newCount) => {
     setTicketCounts((prevCounts) => {
@@ -248,6 +190,7 @@ const NewSchedule = () => {
         availableSeats="20"
         bookedSeats="10"
         updateTicketCounts={updateTicketCounts}
+        updateCost={updateCost}
       />
       <ClassDetails
         seatClass="Second Class"
@@ -256,6 +199,7 @@ const NewSchedule = () => {
         availableSeats="40"
         bookedSeats="26"
         updateTicketCounts={updateTicketCounts}
+        updateCost={updateCost}
       />{" "}
       <ClassDetails
         seatClass="Third Class"
@@ -264,6 +208,8 @@ const NewSchedule = () => {
         availableSeats="58"
         bookedSeats="34"
         updateTicketCounts={updateTicketCounts}
+        updateCost={updateCost}
+
       />
       <Container className="buttonContainer">
         <Row>
@@ -291,7 +237,7 @@ const NewSchedule = () => {
 
             <Row>
               <Col xs={12} sm={4} xl={2}>
-                <p className="ticket-price">3100 LKR</p>
+                <p className="ticket-price">{totalCost}</p>
               </Col>
               <Col xs={12} sm={8}>
                 <p className="cost-para">
